refactor(eslint-rules): split consistent-base-hook into base-hook-signature + base-hook-no-forbidden-runtime
Replaces the single `@nx/workspace-consistent-base-hook` rule with two
focused rules:

- `@nx/workspace-base-hook-signature` (untyped) — enforces the API
  contract for base hooks: 1–2 positional parameters named `props` and
  optional `ref`, with `ref` typed as `React.Ref<...>`. Detects targets
  by name pattern (`use*Base_unstable`) OR by pair detection — when a
  `use<Name>_unstable` exists alongside a co-located or sibling-file
  `use<Name>Base_unstable`, the state hook is also validated.

- `@nx/workspace-base-hook-no-forbidden-runtime` (type-aware) — enforces
  that base hooks do not transitively pull forbidden runtime packages
  (default `tabster`) directly or via configured watched packages
  (default `@fluentui/react-tabster`).

Extracted shared utilities:

- `utils/tracked-imports.ts` — import-specifier metadata and naming.
- `utils/module-resolver.ts` — TS module resolution + bare-spec helpers.
- `utils/transitive-reach.ts` — per-program cached value/all reach
  analysis with cycle handling.
- `utils/base-hook-detector.ts` — name patterns, base/state pairing,
  same-file + sibling-file detection.

Other changes:
- `packages/eslint-plugin/src/internal.js` — register both new rules in
  place of the old one.
- `packages/react-components/react-tags/.../useTagGroup.ts` — suppression
  comment updated to point at the new `base-hook-signature` rule name.
- Fixture dir renamed
  `__fixtures__/consistent-base-hook/` → `__fixtures__/base-hook-no-forbidden-runtime/`.
  Signature fixture lives at `__fixtures__/base-hook-signature/`.

Tests: 76/76 pass (5 suites). vNext lint sweep clean.
