+++ packages/eslint-plugin/src/internal.js
@@ -35,7 +35,8 @@ const __internal = {
           /** @type {import('eslint').Linter.RulesRecord} */
           rules: {
             '@nx/workspace-consistent-callback-type': 'error',
-            '@nx/workspace-consistent-base-hook': 'error',
+            '@nx/workspace-base-hook-signature': 'error',
+            '@nx/workspace-base-hook-no-forbidden-runtime': 'error',
             '@nx/workspace-no-restricted-globals': restrictedGlobals.react,
             '@nx/workspace-no-missing-jsx-pragma': ['error', { runtime: 'automatic' }],
           },
