+++ tools/eslint-rules/rules/__fixtures__/base-hook-no-forbidden-runtime/src/dummy.ts
