+++ tools/eslint-rules/rules/__fixtures__/base-hook-no-forbidden-runtime/stubs/light-helper/index.ts
