--- tools/eslint-rules/rules/__fixtures__/consistent-base-hook/src/components/Sibling/useSiblingBase.ts
@@ -1,7 +1,0 @@
-import * as React from 'react';
-
-// Sibling base hook used to verify cross-file pair detection in the rule's spec.
-// The wrapping state hook lives in `useSibling.ts` next to this file.
-export const useSiblingBase_unstable = (props: { a: number }, ref: React.Ref<HTMLElement>) => {
-  return { props, ref };
-};
