--- tools/eslint-rules/rules/consistent-base-hook.spec.ts
@@ -1,624 +1,0 @@
-import * as path from 'node:path';
-import { RuleTester } from '@typescript-eslint/rule-tester';
-import { rule, RULE_NAME } from './consistent-base-hook';
-
-const FIXTURE_ROOT = path.join(__dirname, '__fixtures__/consistent-base-hook');
-const TYPED_FILENAME = 'src/test.ts';
-const SIBLING_FILENAME = path.join(FIXTURE_ROOT, 'src/components/Sibling/useSibling.ts');
-const ORPHAN_FILENAME = path.join(FIXTURE_ROOT, 'src/components/Orphan/useOrphan.ts');
-
-const typedLanguageOptions = {
-  parserOptions: {
-    project: path.join(FIXTURE_ROOT, 'tsconfig.json'),
-    tsconfigRootDir: FIXTURE_ROOT,
-  },
-};
-
-// ---------------------------------------------------------------------------
-// Parameter-shape checks — no typed linting required.
-// ---------------------------------------------------------------------------
-const ruleTester = new RuleTester();
-
-ruleTester.run(RULE_NAME, rule, {
-  valid: [
-    // Valid base hook: namespace import — `import * as React from 'react'` + `React.Ref<...>`.
-    {
-      code: `
-        import * as React from 'react';
-        export const useThingBase_unstable = (props, ref: React.Ref<HTMLElement>) => {
-          return { props, ref };
-        };
-      `,
-    },
-    // Valid base hook: named import — `import { Ref } from 'react'` + `Ref<...>` (FunctionDeclaration form).
-    {
-      code: `
-        import { Ref } from 'react';
-        export function useThingBase_unstable(props, ref: Ref<HTMLElement>) {
-          return { props, ref };
-        }
-      `,
-    },
-    // Valid base hook: default import — `import React from 'react'` + `React.Ref<...>`.
-    {
-      code: `
-        import React from 'react';
-        export const useThingBase_unstable = (props, ref: React.Ref<HTMLElement>) => {
-          return { props, ref };
-        };
-      `,
-    },
-    // Valid base hook with only \`props\` (ref is optional).
-    {
-      code: `
-        export const useThingBase_unstable = (props) => {
-          return { props };
-        };
-      `,
-    },
-    // Non-base hook is not subject to the param-shape constraint.
-    {
-      code: `
-        export const useThing_unstable = (props, ref, extra) => {
-          return { props, ref, extra };
-        };
-      `,
-    },
-    // Pair detection (same file): a state hook `useThing_unstable` next to its base hook
-    // `useThingBase_unstable` IS subject to the contract. Correct signature passes.
-    {
-      code: `
-        import * as React from 'react';
-        export const useThing_unstable = (props, ref: React.Ref<HTMLElement>) => {
-          return { props, ref };
-        };
-        export const useThingBase_unstable = (props, ref: React.Ref<HTMLElement>) => {
-          return { props, ref };
-        };
-      `,
-    },
-    // Pair detection (no sibling base hook on disk): `useOrphanContextValues_unstable(state)`
-    // is NOT a paired wrapping hook and must NOT be flagged for its non-(props, ref) signature.
-    // The Orphan folder has no `useOrphanContextValuesBase.ts(x)` next to it.
-    {
-      filename: ORPHAN_FILENAME,
-      code: `
-        export function useOrphanContextValues_unstable(state) {
-          return { state };
-        }
-      `,
-    },
-    // Pair detection (sibling file): wrapping state hook lives in `useSibling.ts`, paired with
-    // `useSiblingBase.ts` in the same folder. Correct (props, ref) signature passes.
-    {
-      filename: SIBLING_FILENAME,
-      code: `
-        import * as React from 'react';
-        export const useSibling_unstable = (props, ref: React.Ref<HTMLElement>) => {
-          return { props, ref };
-        };
-      `,
-    },
-    // Identifier with the same local name as a forbidden import alias does not collide via scope analysis.
-    {
-      code: `
-        import * as React from 'react';
-        import { useArrowNavigationGroup } from '@fluentui/react-tabster';
-        export const useThing_unstable = (props, ref: React.Ref<HTMLElement>) => {
-          return useArrowNavigationGroup({});
-        };
-        export const useThingBase_unstable = (props, ref: React.Ref<HTMLElement>) => {
-          const useArrowNavigationGroup = () => 1;
-          return { value: useArrowNavigationGroup() };
-        };
-      `,
-    },
-    // \`keyborg\` is not in the default forbidden runtime list — bindings imported from it are allowed inside base hooks.
-    {
-      code: `
-        import * as React from 'react';
-        import { createKeyborg, KEYBORG_FOCUSIN } from 'keyborg';
-        export const useThingBase_unstable = (props, ref: React.Ref<HTMLElement>) => {
-          return { kb: createKeyborg(window), evt: KEYBORG_FOCUSIN };
-        };
-      `,
-    },
-  ],
-  invalid: [
-    // Too few params (0).
-    {
-      code: `
-        export const useThingBase_unstable = () => ({});
-      `,
-      errors: [{ messageId: 'invalidParamCount', data: { hookName: 'useThingBase_unstable', actual: 0 } }],
-    },
-    // Too many params.
-    {
-      code: `
-        export const useThingBase_unstable = (props, ref, extra) => ({ props, ref, extra });
-      `,
-      errors: [{ messageId: 'invalidParamCount', data: { hookName: 'useThingBase_unstable', actual: 3 } }],
-    },
-    // Wrong param names.
-    {
-      code: `
-        export const useThingBase_unstable = (p, r) => ({ p, r });
-      `,
-      errors: [
-        {
-          messageId: 'invalidParamName',
-          data: { hookName: 'useThingBase_unstable', index: 1, expected: 'props', actual: 'p' },
-        },
-        {
-          messageId: 'invalidParamName',
-          data: { hookName: 'useThingBase_unstable', index: 2, expected: 'ref', actual: 'r' },
-        },
-      ],
-    },
-    // ObjectPattern for \`props\` is not allowed.
-    {
-      code: `
-        import * as React from 'react';
-        export const useThingBase_unstable = ({ a }, ref: React.Ref<HTMLElement>) => ({ a, ref });
-      `,
-      errors: [
-        {
-          messageId: 'invalidParamName',
-          data: { hookName: 'useThingBase_unstable', index: 1, expected: 'props', actual: '{ ... }' },
-        },
-      ],
-    },
-    // \`ref\` parameter without a type annotation.
-    {
-      code: `
-        export const useThingBase_unstable = (props, ref) => ({ props, ref });
-      `,
-      errors: [
-        {
-          messageId: 'invalidRefType',
-          data: { hookName: 'useThingBase_unstable', actual: '<missing type annotation>' },
-        },
-      ],
-    },
-    // \`ref\` parameter typed as something other than React.Ref.
-    {
-      code: `
-        export const useThingBase_unstable = (props, ref: HTMLElement) => ({ props, ref });
-      `,
-      errors: [
-        {
-          messageId: 'invalidRefType',
-          data: { hookName: 'useThingBase_unstable', actual: 'HTMLElement' },
-        },
-      ],
-    },
-    // \`ref\` parameter typed as React.ForwardedRef (must be React.Ref).
-    {
-      code: `
-        export const useThingBase_unstable = (props, ref: React.ForwardedRef<HTMLElement>) => ({ props, ref });
-      `,
-      errors: [
-        {
-          messageId: 'invalidRefType',
-          data: { hookName: 'useThingBase_unstable', actual: 'React.ForwardedRef' },
-        },
-      ],
-    },
-    // \`Ref\` is a locally declared type alias, not imported from react.
-    {
-      code: `
-        type Ref<T> = { current: T | null };
-        export const useThingBase_unstable = (props, ref: Ref<HTMLElement>) => ({ props, ref });
-      `,
-      errors: [
-        {
-          messageId: 'invalidRefType',
-          data: { hookName: 'useThingBase_unstable', actual: 'Ref' },
-        },
-      ],
-    },
-    // \`Ref\` imported from a non-react package is not accepted.
-    {
-      code: `
-        import { Ref } from 'not-react';
-        export const useThingBase_unstable = (props, ref: Ref<HTMLElement>) => ({ props, ref });
-      `,
-      errors: [
-        {
-          messageId: 'invalidRefType',
-          data: { hookName: 'useThingBase_unstable', actual: 'Ref' },
-        },
-      ],
-    },
-    // \`React\` is a locally declared identifier, not the react module namespace.
-    {
-      code: `
-        const React = { Ref: null };
-        export const useThingBase_unstable = (props, ref: React.Ref<HTMLElement>) => ({ props, ref });
-      `,
-      errors: [
-        {
-          messageId: 'invalidRefType',
-          data: { hookName: 'useThingBase_unstable', actual: 'React.Ref' },
-        },
-      ],
-    },
-    // Referencing a watched-package binding inside a base hook without typed services available
-    // surfaces a one-shot `typedServicesUnavailable` diagnostic so the misconfiguration is visible.
-    {
-      code: `
-        import * as React from 'react';
-        import { useArrowNavigationGroup } from '@fluentui/react-tabster';
-        export const useThingBase_unstable = (props, ref: React.Ref<HTMLElement>) => {
-          return useArrowNavigationGroup({});
-        };
-      `,
-      errors: [
-        {
-          messageId: 'typedServicesUnavailable',
-          data: {
-            watchedPackages: '@fluentui/react-tabster',
-            forbiddenRuntimes: 'tabster',
-          },
-        },
-      ],
-    },
-    // Pair detection (same file): wrapping state hook with too many params is flagged because
-    // its sibling base hook in the same file marks it as a paired wrapper.
-    {
-      code: `
-        import * as React from 'react';
-        export const useThing_unstable = (props, ref, extra) => ({ props, ref, extra });
-        export const useThingBase_unstable = (props, ref: React.Ref<HTMLElement>) => ({ props, ref });
-      `,
-      errors: [{ messageId: 'invalidParamCount', data: { hookName: 'useThing_unstable', actual: 3 } }],
-    },
-    // Pair detection (sibling file): wrapping state hook in `useSibling.ts` is paired with
-    // `useSiblingBase.ts` in the same folder. Wrong param names are flagged.
-    {
-      filename: SIBLING_FILENAME,
-      code: `
-        import * as React from 'react';
-        export const useSibling_unstable = (p, r: React.Ref<HTMLElement>) => ({ p, r });
-      `,
-      errors: [
-        {
-          messageId: 'invalidParamName',
-          data: { hookName: 'useSibling_unstable', index: 1, expected: 'props', actual: 'p' },
-        },
-        {
-          messageId: 'invalidParamName',
-          data: { hookName: 'useSibling_unstable', index: 2, expected: 'ref', actual: 'r' },
-        },
-      ],
-    },
-  ],
-});
-
-// ---------------------------------------------------------------------------
-// Forbidden-runtime + transitive-reach checks — require typed linting.
-// ---------------------------------------------------------------------------
-const typedRuleTester = new RuleTester();
-
-const transitiveOptions: readonly [{ watchedPackages: string[]; forbiddenRuntimes: string[] }] = [
-  {
-    watchedPackages: ['watched-pkg'],
-    forbiddenRuntimes: ['heavy-runtime'],
-  },
-];
-
-const transitiveOptionsAllowTypeImports: readonly [
-  { watchedPackages: string[]; forbiddenRuntimes: string[]; allowTypeImports: boolean },
-] = [
-  {
-    watchedPackages: ['watched-pkg'],
-    forbiddenRuntimes: ['heavy-runtime'],
-    allowTypeImports: true,
-  },
-];
-
-typedRuleTester.run(`${RULE_NAME} (typed)`, rule, {
-  valid: [
-    // The defining file of \`useLight\` only reaches \`light-helper\`, not \`heavy-runtime\`.
-    {
-      languageOptions: typedLanguageOptions,
-      filename: TYPED_FILENAME,
-      options: transitiveOptions,
-      code: `
-        import * as React from 'react';
-        import { useLight } from 'watched-pkg';
-        export const useThingBase_unstable = (props: { a: number }, ref: React.Ref<HTMLElement>) => {
-          useLight();
-          return { props, ref };
-        };
-      `,
-    },
-    // Type-only import of a watched-package symbol whose defining file does NOT reach the
-    // forbidden runtime is allowed (`LightOptions` is defined in `light.ts` which only pulls
-    // `light-helper`).
-    {
-      languageOptions: typedLanguageOptions,
-      filename: TYPED_FILENAME,
-      options: transitiveOptions,
-      code: `
-        import * as React from 'react';
-        import type { LightOptions } from 'watched-pkg';
-        export const useThingBase_unstable = (props: LightOptions, ref: React.Ref<HTMLElement>) => {
-          return { props, ref };
-        };
-      `,
-    },
-    // Watched-package import exists but only used by a non-base hook in the same file.
-    {
-      languageOptions: typedLanguageOptions,
-      filename: TYPED_FILENAME,
-      options: transitiveOptions,
-      code: `
-        import * as React from 'react';
-        import { useHeavy } from 'watched-pkg';
-        export const useThingBase_unstable = (props: { a: number }, ref: React.Ref<HTMLElement>) => {
-          return { props, ref };
-        };
-        export const useThing_unstable = (props, ref: React.Ref<HTMLElement>) => {
-          return useHeavy();
-        };
-      `,
-    },
-    // Cyclic re-export graph must not infinite-loop; \`useA\` does not reach heavy-runtime.
-    {
-      languageOptions: typedLanguageOptions,
-      filename: TYPED_FILENAME,
-      options: [
-        {
-          watchedPackages: ['cyclic-pkg'],
-          forbiddenRuntimes: ['heavy-runtime'],
-        },
-      ],
-      code: `
-        import * as React from 'react';
-        import { useA } from 'cyclic-pkg';
-        export const useThingBase_unstable = (props: { a: number }, ref: React.Ref<HTMLElement>) => {
-          return { props, ref, value: useA() };
-        };
-      `,
-    }, // With `allowTypeImports: true`, type-only imports from a forbidden runtime are permitted.
-    {
-      languageOptions: typedLanguageOptions,
-      filename: TYPED_FILENAME,
-      options: transitiveOptionsAllowTypeImports,
-      code: `
-        import * as React from 'react';
-        import type { HeavyOptions } from 'heavy-runtime';
-        export const useThingBase_unstable = (props: HeavyOptions, ref: React.Ref<HTMLElement>) => {
-          return { props, ref };
-        };
-      `,
-    },
-    // With `allowTypeImports: true`, per-specifier type-only import from a forbidden runtime is permitted.
-    {
-      languageOptions: typedLanguageOptions,
-      filename: TYPED_FILENAME,
-      options: transitiveOptionsAllowTypeImports,
-      code: `
-        import * as React from 'react';
-        import { type HeavyOptions } from 'heavy-runtime';
-        export const useThingBase_unstable = (props: HeavyOptions, ref: React.Ref<HTMLElement>) => {
-          return { props, ref };
-        };
-      `,
-    },
-    // Symmetric `allowTypeImports`: also exempts watched-package type-only imports whose defining
-    // module reaches a forbidden runtime (no runtime coupling is possible from a type).
-    {
-      languageOptions: typedLanguageOptions,
-      filename: TYPED_FILENAME,
-      options: transitiveOptionsAllowTypeImports,
-      code: `
-        import * as React from 'react';
-        import type { HeavyType } from 'watched-pkg';
-        export const useThingBase_unstable = (props: HeavyType, ref: React.Ref<HTMLElement>) => {
-          return { props, ref };
-        };
-      `,
-    },
-  ],
-  invalid: [
-    // Direct import from a forbidden-runtime package.
-    {
-      languageOptions: typedLanguageOptions,
-      filename: TYPED_FILENAME,
-      options: transitiveOptions,
-      code: `
-        import * as React from 'react';
-        import { runHeavy } from 'heavy-runtime';
-        export const useThingBase_unstable = (props: { a: number }, ref: React.Ref<HTMLElement>) => {
-          return { props, ref, x: runHeavy() };
-        };
-      `,
-      errors: [
-        {
-          messageId: 'forbiddenRuntimeDirect',
-          data: {
-            hookName: 'useThingBase_unstable',
-            importedName: 'runHeavy',
-            package: 'heavy-runtime',
-          },
-        },
-      ],
-    },
-    // Symbol from watched package whose defining file transitively imports the forbidden runtime.
-    {
-      languageOptions: typedLanguageOptions,
-      filename: TYPED_FILENAME,
-      options: transitiveOptions,
-      code: `
-        import * as React from 'react';
-        import { useHeavy } from 'watched-pkg';
-        export const useThingBase_unstable = (props: { a: number }, ref: React.Ref<HTMLElement>) => {
-          return { props, ref, x: useHeavy() };
-        };
-      `,
-      errors: [
-        {
-          messageId: 'forbiddenRuntimeReach',
-          data: {
-            hookName: 'useThingBase_unstable',
-            importedName: 'useHeavy',
-            package: 'watched-pkg',
-            runtime: 'heavy-runtime',
-            viaFile: 'rules/__fixtures__/consistent-base-hook/stubs/watched-pkg/heavy.ts',
-          },
-        },
-      ],
-    },
-    // Aliased import from a forbidden-runtime package is still flagged on the alias use site.
-    {
-      languageOptions: typedLanguageOptions,
-      filename: TYPED_FILENAME,
-      options: transitiveOptions,
-      code: `
-        import * as React from 'react';
-        import { runHeavy as go } from 'heavy-runtime';
-        export function useThingBase_unstable(props: { a: number }, ref: React.Ref<HTMLElement>) {
-          return go();
-        }
-      `,
-      errors: [
-        {
-          messageId: 'forbiddenRuntimeDirect',
-          data: {
-            hookName: 'useThingBase_unstable',
-            importedName: 'runHeavy',
-            package: 'heavy-runtime',
-          },
-        },
-      ],
-    },
-    // By default, a top-level type-only import from a forbidden runtime is disallowed.
-    {
-      languageOptions: typedLanguageOptions,
-      filename: TYPED_FILENAME,
-      options: transitiveOptions,
-      code: `
-        import * as React from 'react';
-        import type { HeavyOptions } from 'heavy-runtime';
-        export const useThingBase_unstable = (props: HeavyOptions, ref: React.Ref<HTMLElement>) => {
-          return { props, ref };
-        };
-      `,
-      errors: [
-        {
-          messageId: 'forbiddenRuntimeDirect',
-          data: {
-            hookName: 'useThingBase_unstable',
-            importedName: 'HeavyOptions',
-            package: 'heavy-runtime',
-          },
-        },
-      ],
-    },
-    // By default, a per-specifier type-only import from a forbidden runtime is also disallowed.
-    {
-      languageOptions: typedLanguageOptions,
-      filename: TYPED_FILENAME,
-      options: transitiveOptions,
-      code: `
-        import * as React from 'react';
-        import { type HeavyOptions } from 'heavy-runtime';
-        export const useThingBase_unstable = (props: HeavyOptions, ref: React.Ref<HTMLElement>) => {
-          return { props, ref };
-        };
-      `,
-      errors: [
-        {
-          messageId: 'forbiddenRuntimeDirect',
-          data: {
-            hookName: 'useThingBase_unstable',
-            importedName: 'HeavyOptions',
-            package: 'heavy-runtime',
-          },
-        },
-      ],
-    },
-    // Type-leakage through a watched package: a top-level `import type` of a watched-package
-    // symbol whose defining module transitively reaches the forbidden runtime is disallowed
-    // because the type still ties the base hook's public API to the forbidden runtime.
-    {
-      languageOptions: typedLanguageOptions,
-      filename: TYPED_FILENAME,
-      options: transitiveOptions,
-      code: `
-        import * as React from 'react';
-        import type { HeavyType } from 'watched-pkg';
-        export const useThingBase_unstable = (props: HeavyType, ref: React.Ref<HTMLElement>) => {
-          return { props, ref };
-        };
-      `,
-      errors: [
-        {
-          messageId: 'forbiddenRuntimeReach',
-          data: {
-            hookName: 'useThingBase_unstable',
-            importedName: 'HeavyType',
-            package: 'watched-pkg',
-            runtime: 'heavy-runtime',
-            viaFile: 'rules/__fixtures__/consistent-base-hook/stubs/watched-pkg/heavy.ts',
-          },
-        },
-      ],
-    },
-    // Per-specifier `type` modifier variant of the same scenario.
-    {
-      languageOptions: typedLanguageOptions,
-      filename: TYPED_FILENAME,
-      options: transitiveOptions,
-      code: `
-        import * as React from 'react';
-        import { type HeavyType, useLight } from 'watched-pkg';
-        export const useThingBase_unstable = (props: HeavyType, ref: React.Ref<HTMLElement>) => {
-          useLight();
-          return { props, ref };
-        };
-      `,
-      errors: [
-        {
-          messageId: 'forbiddenRuntimeReach',
-          data: {
-            hookName: 'useThingBase_unstable',
-            importedName: 'HeavyType',
-            package: 'watched-pkg',
-            runtime: 'heavy-runtime',
-            viaFile: 'rules/__fixtures__/consistent-base-hook/stubs/watched-pkg/heavy.ts',
-          },
-        },
-      ],
-    },
-    // Indirect type leakage: `HeavyWrapper` is declared in `watched-pkg/index.ts` (not in `heavy.ts`),
-    // but its defining file value-re-exports `./heavy`, so the type-graph reach from `index.ts` still
-    // includes `heavy-runtime`. The base hook surface is therefore tied to the forbidden runtime.
-    {
-      languageOptions: typedLanguageOptions,
-      filename: TYPED_FILENAME,
-      options: transitiveOptions,
-      code: `
-        import * as React from 'react';
-        import type { HeavyWrapper } from 'watched-pkg';
-        export const useThingBase_unstable = (props: HeavyWrapper, ref: React.Ref<HTMLElement>) => {
-          return { props, ref };
-        };
-      `,
-      errors: [
-        {
-          messageId: 'forbiddenRuntimeReach',
-          data: {
-            hookName: 'useThingBase_unstable',
-            importedName: 'HeavyWrapper',
-            package: 'watched-pkg',
-            runtime: 'heavy-runtime',
-            viaFile: 'rules/__fixtures__/consistent-base-hook/stubs/watched-pkg/index.ts',
-          },
-        },
-      ],
-    },
-  ],
-});
